requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pillow>=10.4.0
numpy>=1.26.0
scipy>=1.13.0
scikit-image>=0.24.0
pydantic>=2.7.0
pytest>=8.2.0
httpx>=0.27.0
